fix: adjustments after code review
